[refacto] Create an interval utils
